tests/submission/W3C/domhtml/HTMLStyleElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<style media="screen" type="text/css"></style>
<title>NIST DOM HTML Test - STYLE</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-style-element.html#dom-style-disabled" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>Hello, World.</p>



<script>
var domtest = async_test("The disabled attribute enables/disables the stylesheet.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vdisabled;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("style");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vdisabled = testNode.disabled;

      assert_false(vdisabled);

    domtest.done();
}));

</script>
<p>The disabled attribute enables/disables the stylesheet.            

    Retrieve the disabled attribute and examine its value.</p>
<div id="log"></div>
</body></html>
